super_funsearch/requirements.txt
numpy
scipy
absl-py
torch
tensorboard
tensorboardX
numba
python-dotenv
sentence-transformers